WasmFS: Force-include FS only when FORCE_FILESYSTEM (#19511)
This just applies the principle consistently. The only issues in the test suite
were 3 tests that did manual FS.* operations without setting
FORCE_FILESYSTEM, so add that to them.

As the test updates show this helps with code size (removing unused JS
helps by itself, and also helps metadce remove wasm things that the JS
was pointing to).