WasmFS: Implement FS.cwd (#19452)
This both adds JS API support for it, and internal support in JS libraries that use it
via PATH_FS, such as SDL1.

Also switch the JS API methods to hold on to their allocations rather than return
them to JS and expect JS to call free. This may use more memory in some cases,
but in others will be more efficient, and it is simpler.