WasmFS JS API: Update write to match object returned by open (#19548)
My original PR for write (#19508) passed tests assuming that open returned
an FD. We changed open to return an object in a PR after the write PR (#19542).
test_fs_write failed for write after the open PR was merged in. This PR updates
write to accept an object from open instead of just an fd.