WasmFS: Add a no-FS library, to avoid linking in WasmFS only for printf (#19544)
For WasmFS to be shippable as the default we need to only link it in when needed.
Otherwise we'd regress vs the old FS, which has the SYSCALLS_REQUIRE_FILESYSTEM
optimization. That works by seeing which syscalls are present: if there is no open()
and just write() then we have some printf code but no real filesystem usage. Then
the JS FS is not included at all in the output, and we just add a minimal printf-supporting
shim. That saves a big amount of code size in small programs, and this PR achieves
something similar for WasmFS.

Rather than the emscripten.py logic for SYSCALLS_REQUIRE_FILESYSTEM, this PR
uses library linking. It adds a wasmfs_no_fs library that is linked in right before
WasmFS itself. If all we need are the minimal syscalls for printf, then that library
ends up linked in instead of all of WasmFS.

As shown by test/other/metadce/test_metadce_hello_wasmfs.size, this reduces the
size of a hello world wasm size from 44,622 bytes to just 1,778 (which is almost
the same as without WasmFS, 1,755).

Fixes #19460