Simplify modifyFunction helper function. NFC
The name of the function is not important at any of the call sites since
we only use this helper to modify JS library functions, and JS library
functions get their name set explicitly at the top of
`processLibraryFunction`.

Also, rename the function just in case there are external caller of this
function, this will make it clear they need to update their code.

This change was motivated by #19539 where it helps to avoid worrying
about functions names.