WasmFS: Add wasmfs_flush() and use that to get unflushed content checks working (#19448)
We have assertions on unflushed content, which are important as on the Web
we only print full lines with console.log(), so printf("hello\nworld"); only
prints the first word, and the second waits on a newline before we do a
console.log. The assertions see if there is unflushed content and direct the
user to how to fix this potentially confusing problem.

For the assertion to work, we need to flush not only musl's libc buffers but also
our own buffering before console.log. In the JS FS that is done in JS code. For
WasmFS, we need to flush them in C++, which this PR adds.

This makes that flushing method a public API, since it seems like a generally
useful thing.