[WasmFS] Eagerly read directory entries when opening directories (#19556)
This ensures that subsequent getdents calls see a stable set of entries, even as
entries are removed or inserted after the open. Although on its face this sounds
like it would be incorrect behavior, it is in fact allowed by the spec and
solves a bug where deleting entries while iterating through them would cause
some entries to be skipped.

Fixes #19106.