WasmFS: Include all WasmFS JS code when INCLUDE_FULL_LIBRARY (#19555)
The issue is that only FORCE_FILESYSTEM causes the full JS API code to be present,
but other places refer to FS.cwd() (PATH_FS). What we kind of want is for PATH_FS
to cause FORCE_FILESYSTEM to be set, but it's too late for that in the compilation
process.