Enable `--proxy-to-worker` under node.
At least enough to run hello world.  The idea here is not so much that
folks really want use `--proxy-to-worker` under node, but more for
testing, so we can run tests where the module is instantiated on a
worker.

In particular we can write tests for issues such as #19577 that run
under node.

Also, because `proxyClient.js` now goes through the preprocessor we can
use the normal `#include` mechanism to include `IDBStore.js`.