Fix SDL1 deps (#19610)
FS is only used in one specific method, so put the dep there. That avoids a
WasmFS closure error on including FS without enough WasmFS code (which
should not happen in practice, as that method loads data from a file - so
WasmFS would be present to support those file operatoins).

Also intArrayFromString is used and was not declared.