WasmFS: Fix file packager use after startup (#19615)
During startup if we run file packager code then the createDataFile etc. calls
are turned into code that caches the info for later, as we can't call into
WasmFS during startup. Then when WasmFS is running, it reads the cached
info and creates the files. But we also need to handle the case where file
packager code is run after startup, which is easier to support even if it
is more rare - just create the file or directory.