WasmFS: Fix FS.cwd() usage without FORCE_FILESYSTEM (#19616)
The linking of FS.cwd is pretty awkward, unfortunately, but without adding too
much more complexity this PR makes it work in another case, when
FORCE_FILESYSTEM is not set. I am hoping this is the end of the complexity
here. See more details in the big comment.

The enabled test is an example of a case that failed before.

This PR also allows us to remove the ugly INCLUDE_FULL_LIBRARY hack in this
code.