WasmFS: Fix and add deps (#19630)
Use the same method on readfile as on cwd. That works better for a method
inside a big object like FS. And it really makes a lot of sense the more I get
used to it: when the wasm code is there, expose the JS method that calls it,
and not otherwise; that avoids all possible problems (aside from code not
including the wasm code when it needs to - but that would be a bug in the
other code).

This gets us much closer to being able to run a closure+FULL_LIBRARY test.