Improve behavior of interval timers in setitimer
Alarm firing can be delayed (i.e. occur after the time it was due).
When this happens the next timer is due 'interval' ms after the
previous scheduled time (which in case of delay is sooner then
'interval' ms after the time it actually fired).

These delays (seen when tests run under load) were causing flakiness
in test_itimer on bots; in particularly the test_sequence test.

There are 2 things that can still go wrong with test_itimer:
1) A single timer event can be delayed long enough to cause test_oneoff
to fail (this happens if the delay is > 100ms in the current version of the
test). I have seen this recently on the bots as well. This PR increases the
sleep time while waiting for the signal, but it could still fail under heavy
load.

2) If a delay in timer firing lasts longer than the interval itself, then
a timer event will be missed entirely. This could also cause the test to
fail (although I haven't seen it in practice).