Remove receiveI64ParamAsI32s helper function (#19701)
I'm working on a tools that automatically converts i64 to i53 without
any extra macros or helpers in the source code.

This way to receiving an i64 was an outlier and removing it helps
towards my goal.  The other problem with this method is that it
doesn't make a lot of sense in WASM_BIGINT is enabled since it needless
splits the incoming argument into two.