Remove EMCC_SAVE_OPT_TEMP environment variable
The files in questions are already save by when `EMCC_DEBUG` (or just
`EMCC_DEBUG_SAVE`) are set.  Having a third options seems rather
pointless.

Furthermore `EMCC_SAVE_OPT_TEMP` doesn't actually work on its own since
it save files into the emscripten temp directory, which is completely
removed unless one of the above options I used.

In addition this code is currently broken due to the expression:

  `int(saved.replace('input', '').replace('.txt', ''))`

It seems like the above code was assuming the intput was of the form
`include<N>.txt` but that is not the case since `saved` is declared on
the line above as:

  `saved = 'save_' + os.path.basename(filename)`

Also remove an unused function from `tools/shared.py`