Create launcher scripts as part of `bootstrap.py`
Followup to #19736

This allows us to do is delete all the checked in launcher files from
the top level directory.  This in turn means we can create just the
`.bat` file launchers on windows and just the `.sh` launchers on
non-windows.  The is a small step towards having a much cleaner
top-level directory.

Note that I have left the launchers for emcc checked-in (at least for
the transition period) so that new users will see the error message and
know to run `bootstrap.py` if they run `emcc` on a fresh checkout.