Remove unnecessary try/catch in instantiateSync. NFC (#19784)
The catch only existed to handle cases where `IMPORTED_MEMORY` was set
and `INITIAL_MEMORY` was overridden at runtime to a value that was too
low.  This is such a rare setup that I don't believe it warrents this
special case.  In the past, setting `INITIAL_MEMORY` at runtime was more
common, but since `IMPORTED_MEMORY` now defaults to false
`INITIAL_MEMORY` is not valid at all for most users.

Also, the default error message from the runtime seems seems clear
enough.

Also, simply and if-after-else.

Also, rename getBinary to getBinarySync, so it explicitly contrasts with
the other getBinaryPromise.