Skip `fetch()` code under `SINGLE_FILE` (#19789)
This change also removes the `test_base64_atob_fallback` test which was
supposed to be testing that "base64 utils work in browser with no native
atob function".  However, the base64 fallback was removed by default
in #19606.  This test was passing accidentally because, as it happens,
XMLHttpRequest works with data URLs and the code path being removed
in this change was going through XHR request rather then "atob". So the
test was already not testing what it claimed to be, and this change
just highlighted that fact.

Fixes: #19746