Clarify comments on Emscripten EH-based settings
The original intention was to gather all Emscripten-EH settings in one
place and put a single comment block saying "The below three only apply
to Emscripten EH", but when you search for a setting you just do a
keyword search and it is not easy to look up to see this comment. So
this makes it clear that a certain option only applies to Emscripten EH
in by putting a comment in each setting. Also move
`DISABLE_EXCEPTION_THROWING` up so that all three Emscripten-EH-only
settings are clumped together.