Don't generate function wrappers for exported tags. NFC
Back in #17064 we added tags to `metadata.exports`, which previously
contained only functions.  However, emscripten.py makes wrapper functions
for each element of this list which is not desirable for tag exports.

This change makes things more explicit by calling this
`function_exports` and creating a new `metadata.all_exports`.

We can also now be more precise when creating the metadce graph and
include all exports (except wasm globals which are just constants),
which in turn allows us to remove the hardcoded handling for the table
and memory export.

This should be a minor code size saving for wasm-exceptions-using
programs since we no longer generate the needless function wrapper for
the tag.