Deprecate `EM_LOG_FUNC_PARAMS` flag to emscripten_log/emscripten_get_callstack
Supporting this flag requires using `arguments.callee` which is
deprecated, and won't work if any of functions on the callstack are in
strict mode (or are arrow function, which implies strict mode).

Why any frame in the calling stack is in strict mode this call will
fail with:

```
TypeError: 'caller', 'callee', and 'arguments' properties may not be accessed on strict mode functions or the arguments objects for calls to them
    at traverseStack (emscripten_log.js:941:30)
```

This effects of these flag seem to not actually be tested, even though
it is used in `test_emscripten_log`.