Don't enable USE_GLFW by default
This changes the default behaviour to match that of STRICT mode
and MINIMAL_RUNTIME node.

We made a similar change for USE_SDL back in #18443.

As part of this change I also updated the code that maps `-l` flags
to JS libraries.  In some cases we also want these to set the
corresponding `-sUSE_XXX` setting.