Fix EM_CALLBACK_THREAD_CONTEXT_CALLING_THREAD when used on the main thread.
Back in #19691 I tried to fix EM_CALLBACK_THREAD_CONTEXT_CALLING_THREAD
but I broken the case where it was used in pthread applications but
on the main thread itself.

We were not previously testing this case because most of our testing
was don't using PROXY_TO_PTHREAD, so the calls we originating from
a non-main thread.

Fixes: #19921