Rename internal JS proxying function. NFC (#19945)
I renamed the existing `_emscripten_run_in_main_runtime_thread_js` to
the simpler `_emscripten_run_in_main_thread_js` to match other things
such as `$proxyToMainThread`.

I don't believe is ambiguous. There is no way this function could be
referring to the main browser thread and not the main runtime thread
since if the main browser thread ever exists in the system it is by
definition also the main runtime thread.

This change is laying the groundwork for a larger change which to move
JS function proxying over to the new proxying system.