Fix for async cancellation during __timedwait. NFC
Emscripten doesn't have real async cancellation but we don't want
__timedwait_cp to return the thread has been canceled while while its
running.

This fixes a problem in the test_pthread_setcanceltype_1_1 test which
was not being correctly reported but shows up as a real problem with
the switch to the new proxying system in #19947.

Specifically, Without this change, a thread this async canceled during
pthread_mutex_lock call will return ECANCELED, but instead it should
never return and thread should be canceled, which is the behaviour
after this change.