Add new internal setting: MINIFY_WHITESPACE (#19961)
This centralizes the control of JS whitespace minification behind a
single setting.  We have had a feature request to be able to disable
this minification so the output can be passed to closure compiler after
link time.

This change doesn't yet allow the user to directly control this setting.
We can consider that part as a followup.

The `test_emcc_output_worker_mjs` test expectation changed because
we now minify the worker.js file based on the same set of criteria as
the main js file.  Previously these were independent and had slightly
different criteria.