Simplify sourcemap generation, and append `sourceMappingURL` section with objcopy (#19890)
This PR simplifies the logic for handling the debug info required for source maps. In particular if debug info will not be needed later, it strips the DWARF earlier than previously (now in the finalize_wasm phase immediately after generating the source map and before finalization). It also uses objcopy rather than Binaryen to add the sourceMappingURL section that points to the source map. This is faster (in cases where Binaryen is not otherwise needed) and means that source map generation works even when using -sERROR_ON_WASM_CHANGES_AFTER_LINK.

Depends on https://reviews.llvm.org/D155535


Co-authored-by: Heejin Ahn <aheejin@gmail.com>