WasmFS: Fix browser pthread asan test reporting (#20005)
These tests work as follows:

* Listen for the print to stderr of the ASan output, check it is right, and report the result
  to the server.
* The reported result sets reportResultToServer = true.
* When we exit from main, we see that true and do not report the exit code from main
  (if we did, we'd error on excessive logging).

That does not work in WasmFS because the print to stderr happens on the pthread
(because we do not proxy every single stdio operation to the main thread constantly).

As a result, we set reportResultToServer = true on one thread, but when main exits
it checks it on another thread, and ends up reporting a second result which errored.

To fix that, let tests that report results themselves set a flag that prevents reporting
the result of main. A downside to this is that if the normal reporting fails, the test will
hang rather than error, but I don't have a good idea of how to avoid that.