WasmFS: malloc-related docs and test fixes (#20004)
Document WasmFS differences vs the old FS, in particular, that -sMALLOC=none does not
work at all.

Also some minor test fixes and cleanups. In test_no_nuthin I tested with emcc_args and
it uncovered an existing annoyance with closure and -sWASM=0 which I filed a bug for +
added a workaround (-Wno-closure).