Remove use of `cgi` module from emrun.py (#20053)
The cgi module is due for removal from python and in recent versions
generates a warning message on import.  Furthermore the only usage of
this module was to call `cgi.escape` in `format_html` and it turns out
that `cgi.escape` does not exist in the `cgi` module, and least not
int python 3.8 or above.

Working backwards I think we can just remove the `--log_html` option
which was the only way to enable `format_html`.