fix `deleted val` problem caused by move assignment to itself
The implementation of `val& operator=(val&& v)`
should not directly call `_emval_decref(handle)`.
According to the semantics of move & assignment,
the input argument(v) need to be reset first,
and then the internal handle is judged to be
a valid handle before calling `_emval_decref`.
This avoids the loss of reference problem that
occurs when moving the assignment to itself (this).