[Wasm64] Fix many remaining issues and add wasm64_4gb test configuration (#20077)
One notable change here the change to node `fs.syncRead/syncWrite`
calls.  Previously we were doing `Buffer.from(buffer.buffer)` but this
was running into the v8 limit on the size of of a typed array:

```
RangeError: Invalid typed array length: 4404019200
```

The new code just makes a small array view of the range being written to
or read from.