Update libcxx/libcxxabi to llvm 16.0.6
This only include readme changes and the version number change in
`libcxx/include/__config`. This is not a terribly meaningful update, but
I'm trying to update libunwind based on llvm 16.0.6 and it would be
simpler if all libraries are based on the same version.

Fixed tag -> branch in `readme.txt`s given that we don't use tags in
https://github.com/emscripten-core/llvm-project.