Move some FS methods to FS_foo notation, allowing WasmFS to use them from JS APIs (#20098)
FS.createDataFile, FS.mkdirTree, FS.unlink are all now usable as FS_foo, that is,
without the FS object being included. For the JS FS this is just a slightly odd
refactoring, but for WasmFS it means we can use dependencies on a per-function
basis, both for including them, and for including what they need.

Also in WasmFS do the same for FS.create, FS.mknod, FS.mkdir, FS.writeFile
as those are dependencies of the above.

This may have a slight code size disadvantage in builds that include the FS object
anyhow, but closure should be able to remove that anyhow. For the JS FS, FS is
there all the time so we don't see any benefits really, but for WasmFS this allows us
to use filesystem operations from JS APIs without bringing in all of FS (and without
all the wasm code it calls into). WasmFS builds that do include FS - which is builds
that have FORCE_FILESYSTEM enabled - may increase in size slightly, but those
builds are rarer and large anyhow due to FS.

With this, emscripten_async_wget can be converted to work in WasmFS. To do so
we switch from FS.foo to FS_foo notation and add the proper deps, and then
everything just works.

Implements the ideas discussed in the later part of #20022