embind: Support strings for names of FunctionDefinitions for TS generation.
`_embind_register_class_constructor` was calling `createFunctionDefinition`
with a string instead of a ptr to a string on the heap. This caused
`readLatinString` to try and construct a string from the whole heap. On
lower memory limits v8 handled this fine, but with higher limits the string
became too large. The test still passed with the bug since the constructor
name is not actually used in the .d.ts output file.