Change _NOEXCEPT back to throw() (#20170)
This changes `_NOEXCEPT` back to `throw()` to follow the upstream
libcxxabi. I'm actually not sure why I changed `throw()` to `_NOEXCEPT`
in #10787 3 years ago, given that we were already handling `throw()` the
same way as `noexcept`:
https://reviews.llvm.org/D79655?vs=263218&id=264335#2039395
https://github.com/llvm/llvm-project/blob/b86d3cbc1288b4d0799746c82611d17888413aec/clang/lib/CodeGen/CGException.cpp#L498-L509
https://github.com/llvm/llvm-project/blob/b86d3cbc1288b4d0799746c82611d17888413aec/clang/lib/CodeGen/CGException.cpp#L602-L609

The upstream libcxxabi inconsistently uses both `noexcept` and
`throw()`, but given that we treat them the same, I think it's better to
minimize the difference between our port and the upstream, especially
now we are trying to upstream the library.