assign $$ is ClassHandle constructor
It appears in some cases `$$` is readonly and will result in a runtime error attempting to do the otherwise no-op assignment meant to help the jscompiler. Instead, we can assign the property in the constructor to undefined.

The error I'm seeing is a runtime error `TypeError: Cannot assign to read only property '$$' of object '[object Object]`. This seems to indicate that `$$` is readonly, but I can't find a call to `Object.freeze` or where `writable` would be set to `false`.