Optimise handleAsync for JSPI
With current handling, in JSPI mode handleAsync was doing unnecessary double wrapping-unwrapping of promises: invoke startAsync, it returns the promise, convert it to callback-style function with wakeUp, pass to handleSleep, handleSleep takes the callback and converts it back to another promise.

For JSPI Promises are native, so it makes more sense to invert the relationsheep between those 2 APIs and make handleAsync just await the original promise, while handleSleep would be the only one converting between callback and promise styles.