Rename the `exports` variable in preamble.js to `instExports` (#20231)
This avoids a bug (or limitation) in Google's Closure compiler that does not
allow any variable named exports to be reassigned. Without this change, some
build configurations result in an error message like this:

[JSC_EXPORT_NOT_AT_MODULE_SCOPE] Exports must be at the top-level of a module  543|
     exports = Asyncify.instrumentWasmExports(exports);  

Since the exports variable used in the preamble is not related to Closure modules,
simply renaming the variable does the trick.