Use `int` to represent webidl `long` type
The webidl `long` type is defined to be a 32-bit type:

https://webidl.spec.whatwg.org/#idl-long
https://webidl.spec.whatwg.org/#idl-unsigned-long

Using `long` on the native side works find for wasm32 but breaks
under wasm64 where `long` is 64-bit.  Using `int` instead means the
type is the same under wasm32 and wasm64 which more accurately
represents the html5/webidl type.

See #20276