Fix pthread_detach such that it only updates the state once
This was causing spurious "Blocking on the main thread is very
dangerous", because if the first a_cas succeeded the second one would
always fail, leading to __pthread_join being called.  `__pthread_join`
is mostly harmless if the thread is already in a DT_DETACHED state, but
it does/did have the side effect of printing this warning.

Fixes: #20224