embind: Also build program normally when using `--embind-emit-tsd`.
Changes the behavior `--embind-emit-tsd` so that the program is built
normally, but the TypeScript definition generation also happens as a
side affect. This is implemented by generating the program normally until
`emscript` phase when a separate "emscript" like phase is run to create
the JS output file, but with `embind_ts.js` replacing `embind.js`. The
instrumented version is then run in node like before. After the TS definitions
are created, the settings and options are restored and compiling finishes
as it normally would.

The test for options that conflict with embind-emit-tsd is now a separate
test since the original test needs to be run and work.