[test] Remove test_noexitruntime. NFC
I'm can't see anything that this test is actually testing.  This
test was added back in 5e633806, but without any explanation why its
useful or what its testing.

`noExitRuntime` is the default so the destructor in this test is
never run, regardless of which EM_ASM block runs.

Furthermore even if the destructor or does run, it would add its error
to the end of stdout, which would not be caught by the check which
(which ignore trailing lines in the output).