[WasmFS] Add an async constructor for the OPFS backend
When pthreads are enabled, the existing OPFS backend constructor synchronously
waits for a worker thread to start up and become ready to receive work. Since
the process of starting the worker thread requires the main thread to return to
its event loop, calling the constructor on the main thread would deadlock. The
previous workaround was to simply not call the constructor on the main thread.

As a more elegant solution, add an async version of the constructor that returns
an em_promise_t that will be fulfilled with the backend handle once the worker
thread has started up. This async constructor can be safely called from the main
thread; the promise simply won't be fulfilled until the main thread gets a
chance to return to its event loop.

Internally, the new constructor works by having the worker thread asynchronously
proxy a task back to the constructing thread when it is ready to receive work.
The task resolves a promise on the constructing thread, which leads to the
user-visible promise getting fulfilled with the value of the backend handle.

Add tests using the new constructor in a fully asynchronous way using pthreads
or JSPI as well as tests that use JSPI to await the result of the constructor
with and without pthreads enabled.