Remove try/catch from intArrayFromBase64. NFC (#20375)
It seems like the error is strictly worse than just letting the
exception propagate out.  For example, we have a report of someone
hitting this error but I can't tell what the root cause was:
https://github.com/emscripten-core/emscripten/issues/20349

This try/catch and this error message was part of the original
function back #5296 but I don't see any reason for them.