[test] Enable entire browser test suite to run in wasm64 mode
This creates a new CI runner/bot that runs the same subset of browser
tests that we run today in wasm64 mode.  This should speedd up CI as
it splits the chrome and firefox tests between two different runners.

However, it makes it possible to run the entire browser test suite in
this mode if/when we want to do that.