Update libunwind CMakeLists.txt to LLVM 16.0.6
This was missing in #20088. I guess this was missing because I did that
update half-manually, before learning to use the scripts in
https://github.com/emscripten-core/emscripten/blob/main/docs/process.md#updating-the-llvm-libraries.
This was discovered while I'm updating libraries to LLVM 17, which
caused a large number of conflicts in this file, even though we don't
use this file and didn't make any changes.