Add `bootstrap.py` script which takes care of post-checkout tasks (#19736)
This script is only present is git checkout of emscripten and not in
emsdk or other packaged versions.

When run with no arguments this script ensures that all the
post-checkout steps required by emscripten have been run and are
up-to-date with the current checkout.  It does this by comparing
`.stamp` files in the `out` directory with the timestamps of inputs
files such as `package.json`.

When `emcc` is run, and bootstrap.py is present it is run in `check`
mode which will cause the compiler to error if the one or more of the
bootstrap commands needs to be run (e.g. if package.json was modified).