Avoid trampolines in emmalloc (#20496)
Rather than define malloc as a small function that calls emmalloc_malloc, just make it
an alias, etc.

In an optimized build the small functions get inlined away, but we can save that work,
and also it can have downsides (e.g. they have different function pointers, which
aliases do not, see discussion in #20487).

For that to work, we need to make the aliases weak, to match the fact that the
trampolines were weak.