Refactor other.test_embind. NFC
This single test was very slow (~2 min) because it was doing its own parametrisation inside test-level parametrisation.

I changed parametrisation helper to allow permutations if it's applied multiple times to the same function, and refactored other.test_embind to use that so that those subtests are parallelised and can be called individually.