Fix LTO runs of test_nothrow_new_nogrow
This test was recently added in #20154 but has been failing on the
emscripten-releases waterfall in thinlto modes.

I'm not sure exactly why, but it seems that perhaps when the value
of `data` is not actually thinLTO simple removes the call the
new/malloc?  I'm not sure why this wasn't also failing under full LTO.