Fix Ninja build on Windows
`:` and ` ` are special from Ninja's PoV and need to be escaped with `$`, while `\` should be just converted to `/`.

See https://github.com/emscripten-core/emscripten/issues/19079#issuecomment-1783151627.

I tested that this works on my Windows machine but not sure how to (and whether worth it) to test this on CI as well.