Set PSTL mode to serial
PSTL was integrated in libc++ recently and we have to choose one out of
three PSTL modes in many places, one of them being
https://github.com/llvm/llvm-project/blob/373c343a77a7afaa07179db1754a52b620dfaf2e/libcxx/include/__algorithm/pstl_backends/cpu_backends/backend.h#L15-L23.

This is supposed to set in `__config_site`, which is supposed to be
generated from
https://github.com/llvm/llvm-project/blob/main/libcxx/include/__config_site.in#L33-L36
from
https://github.com/llvm/llvm-project/blob/main/libcxx/CMakeLists.txt.
Given that we don't use CMake system in Emscripten, we need to hard-code
this in this file. Note that this cannot be solved by just adding
`-D_LIBCPP_PSTL_CPU_BACKEND_SERIAL` in `libcxx`'s `system_lib.py` flags,
because this is needed in header files that are included in user code
as well.