Exclude unnecessary files
compiler-rt's `crtbegin.c` and `crtend.c` were originally in
`compiler-rt/lib/crt`, the directory we don't maintain, and moved to
`compiler-rt/lib/builtins` recently. So they were included in the
`excludes` list.

libcxx's `libdispatch.cpp` is related to PSTL's `LIBDISPATCH` mode,
whatever that is, which doesn't seem to be necessary for us.