Refactor line_endings.py and apply --output-eol to worker.js. NFC
Most of the code in this file is used only for testing and is not part
of emscripten proper, so move that code into test/line_endings.py

Move the remaining utility functions to tools/utils.py.

Update `write_file` utility so it can write a file with the correct
line endings. This simplifies the callers who want to write a file
with specific line endings.