fixed enter/exit fullscreen
The issue is that requestFullscreen (resp. exitFullscreen) is an asynchronous call which can fail: as a result, the code must handle errors using the Promise mechanism.

- now requestFullscreen returns a promise (since it was returning nothing before, any calling code should not be affected)
- now exitFullscreen returns a promise or null (was returning a boolean so the SDL code
  if (Browser.exitFullscreen()) {
      return 1;
    }
  is still valid)
- requestFullscreen handles the error first and call fullscreenChange() which has the effect of removing the div that was set
- requestFullscreen then propagates the error so that the calling code can handle it as well
- glfw detects the error and act appropriately (similarly for exitFullscreen)

Note that this check in contains a bunch of logging to trace what is going on since the asynchronicity makes it hard to trace. These logs will be removed prior to submitting the PR.