Fix test_embind_2's use of pthreads (#20646)
The test was parameterized as having a pthreads and non-pthreads mode, but the
decorator was on the entire test, making it all run as pthreads, which meant the
non-pthreads case was not really tested.

It seems simplest to just refocus the test as a single test, and use pthreads there
for coverage. I don't think we are losing anything of value that way, and it's
simpler than adding logic to support both modes in one test.