Call pthread_attr_getdetachstate after pthread_attr_init
In LLVM 16, in `pthread_create` LSan interceptor, if `attr` is NULL, it
calls `pthread_attr_init` and initializes the `attr` with it, and then
calls `pthread_attr_getdetachstate`:
https://github.com/llvm/llvm-project/blob/7cbf1a2591520c2491aa35339f227775f4d3adf6/compiler-rt/lib/lsan/lsan_interceptors.cpp#L450-L456

In Emscripten, #15099 changes the `if` condition so that even if `attr`
is not NULL, if it is `__ATTRP_C11_THREAD`, we call `pthread_attr_init`.
`__ATTRP_C11_THREAD` looks like something used from musl, and is defined
as -1.
https://github.com/emscripten-core/emscripten/blob/5ce75b8828e3f50494c956d42f2bac301e41253b/system/lib/compiler-rt/lib/lsan/lsan_interceptors.cpp#L465

In LLVM 17, somehow the order of the `pthread_attr_init` and
`pthread_attr_getdetachstate` has swapped:
https://github.com/llvm/llvm-project/blob/309d55140c46384b6de7a7573206cbeba3f7077f/compiler-rt/lib/lsan/lsan_interceptors.cpp#L444-L453
So we don't get to call `pthread_attr_init` before calling
`pthread_attr_getdetachstate`. Even if the new code calls
`pthread_attr_getdetachstate` only when `attr` is not NULL, in our case
it didn't help because our `attr` was not NULL but `__ATTRP_C11_THREAD`.

This swaps the code order back to what it was in LLVM 16. This is
necessary to pass `lsan.test_pthread_c11_threads*`.

Drive-by fix: This also guards
`if (!attr || attr == __ATTRP_C11_THREAD)` condition with
`SANITIZER_EMSCRIPTEN`, which was an Emscripten-specific fix added
before.