embind: Add ahead of time JS glue code generation.
Adds a new setting `-sEMBIND_AOT` that will create all the glue code
invoker functions during compilation. This uses the same framework as the
TypeScript generation code and creates an instrumented version of embind
that is run in node to generate the glue code.

This tries to reuse as much of the `craftInovkerFunction` code that the
current JIT mode uses, so it will remain functionally equivalent.