Remove dependency on assert from SAFE_HEAP. NFC
As of #20592, `assert` is no longer defined in release builds in STRICT
mode, but its still possible to use SAFE_HEAP in this configuration.
Instead of complicating the definition of `assert` to match
`preamble_minimal.js` this change instead removed the dependency of
SAFE_HEAP on the assert function.

This fixes the `core2s.test_module_wasm_memory` which is currently
failing on the emscripten-releases tree.