Remove limits of ES6 features in standard library code
When transpiling to ES5 use closure with `SIMPLE_OPTIMIZATIONS` rather
then `WHITESPACE_ONLY`.  This means that polyfills cant be included
as needed and removes the limits on what ES6 features we can use.

The downside of this is that is slows down builds for ES5 users but
this seems like a reasonable tradeoff.

Fixes: #11984