Simplify clang vs clang++ selection process when compiling sources
Base the choice solely on whether `emcc` or `em++` was used on the
command line.  This means that the handling C vs C++ input files will
now simply match that of the underlying compiler.

When running `clang` there is automatic language selection at the lower
level such that that `clang -c file.cc` will run with `-x c++`
internally and `clang -c file.c` will run with `-x c` internally.

The situation with `clang++` (and now `em++`) is a bit different in
that running `clang++` on `.c` input file not automatically use C mode
and will generate a warning:

```
clang: warning: treating 'c' input as 'c++' when in C++ mode, this behavior is deprecated [-Wdeprecated]
```

Previously we were hiding the warning but using `clang` internally even
if the user ran with `emc++`.